+++ NEWS.md
@@ -1,3 +1,19 @@
+# shinyChakraUI 1.1.0
+
+* Added the range slider (`chakraRangeSlider`), provided by the new version of 
+**Chakra UI**.
+
+* Dependency to the 'fontawesome' package. Fontawesome icons can be used on 
+the thumbs of the sliders.
+
+* Added the Super Tiny Icons.
+
+* Added the ionicons.
+
+* Added the pin input widget (`chakraPinInput`).
+
+
+
 # shinyChakraUI 1.0.0
 
 * First release.